+++ packages/manager/src/features/linodes/LinodesLanding/LinodeRow/LinodeRow.tsx
@@ -9,30 +9,34 @@ import * as React from 'react';
 import { Link } from 'react-router-dom';
 import { compose } from 'recompose';
 import Flag from 'src/assets/icons/flag.svg';
+import Hidden from 'src/components/core/Hidden';
 import Tooltip from 'src/components/core/Tooltip';
+import Typography from 'src/components/core/Typography';
 import HelpIcon from 'src/components/HelpIcon';
-import TableCell from 'src/components/TableCell';
-import TableRow from 'src/components/TableRow';
+import StatusIcon from 'src/components/StatusIcon';
+import TableCell from 'src/components/TableCell/TableCell_CMR';
+import TableRow from 'src/components/TableRow/TableRow_CMR';
 import { Action } from 'src/features/linodes/PowerActionsDialogOrDrawer';
-import { linodeInTransition } from 'src/features/linodes/transitions';
+import {
+  getProgressOrDefault,
+  linodeInTransition,
+  transitionText
+} from 'src/features/linodes/transitions';
 import { DialogType } from 'src/features/linodes/types';
-import useFlags from 'src/hooks/useFlags';
-import { capitalize } from 'src/utilities/capitalize';
+import { capitalize, capitalizeAllWords } from 'src/utilities/capitalize';
 import { linodeMaintenanceWindowString } from '../../utilities';
 import hasMutationAvailable, {
   HasMutationAvailable
 } from '../hasMutationAvailable';
 import IPAddress from '../IPAddress';
 import LinodeActionMenu from '../LinodeActionMenu';
-import LinodeActionMenu_CMR from '../LinodeActionMenu_CMR';
 import RegionIndicator from '../RegionIndicator';
 import { parseMaintenanceStartTime } from '../utils';
 import withNotifications, { WithNotifications } from '../withNotifications';
 import withRecentEvent, { WithRecentEvent } from '../withRecentEvent';
-import styled, { StyleProps } from './LinodeRow.style';
 import LinodeRowBackupCell from './LinodeRowBackupCell';
 import LinodeRowHeadCell from './LinodeRowHeadCell';
-import LinodeRowLoading from './LinodeRowLoading';
+import styled, { StyleProps } from './LinodeRow.style';
 
 interface Props {
   backups: LinodeBackups;
@@ -48,24 +52,28 @@ interface Props {
   vcpus: number;
   status: LinodeStatus;
   displayStatus: string;
-  vlanIP?: string;
   type: null | string;
   tags: string[];
   mostRecentBackup: string | null;
-  openDeleteDialog: (linodeID: number, linodeLabel: string) => void;
-  openPowerActionDialog: (
-    bootAction: Action,
+  vlanIP?: string;
+  isVLAN?: boolean;
+  openTagDrawer: (
     linodeID: number,
     linodeLabel: string,
-    linodeConfigs: Config[]
+    tags: string[]
   ) => void;
   openDialog: (
     type: DialogType,
     linodeID: number,
     linodeLabel?: string
   ) => void;
-  // Including for type matching with CMR; not used.
-  openTagDrawer?: any;
+  openPowerActionDialog: (
+    bootAction: Action,
+    linodeID: number,
+    linodeLabel: string,
+    linodeConfigs: Config[]
+  ) => void;
+  openNotificationDrawer: () => void;
 }
 
 export type CombinedProps = Props &
@@ -86,25 +94,25 @@ export const LinodeRow: React.FC<CombinedProps> = props => {
     region,
     status,
     displayStatus,
-    tags,
     mostRecentBackup,
     disk,
     vcpus,
     memory,
     type,
+    tags,
     image,
+    vlanIP,
     // other props
     classes,
     linodeNotifications,
     openDialog,
-    openDeleteDialog,
     openPowerActionDialog,
+    openNotificationDrawer,
     // displayType, @todo use for M3-2059
     recentEvent,
-    mutationAvailable
+    mutationAvailable,
+    isVLAN
   } = props;
-
-  const flags = useFlags();
 
   const loading = linodeInTransition(status, recentEvent);
   const dateTime = parseMaintenanceStartTime(maintenanceStartTime).split(' ');
@@ -112,11 +120,22 @@ export const LinodeRow: React.FC<CombinedProps> = props => {
   const MaintenanceText = () => {
     return (
       <>
-        For more information, please see your{' '}
-        <Link to="/support/tickets?type=open">open support tickets.</Link>
+        Maintenance is scheduled for{' '}
+        {linodeMaintenanceWindowString(dateTime[0], dateTime[1])}. For more
+        information, please see your{' '}
+        <Link className={classes.statusLink} to="/support/tickets?type=open">
+          open support tickets.
+        </Link>
       </>
     );
   };
+
+  const iconStatus =
+    status === 'running'
+      ? 'active'
+      : ['offline', 'stopped'].includes(status)
+      ? 'inactive'
+      : 'other';
 
   const headCell = (
     <LinodeRowHeadCell
@@ -138,108 +157,115 @@ export const LinodeRow: React.FC<CombinedProps> = props => {
       memory={memory}
       image={image}
       maintenance={maintenanceStartTime}
+      isVLAN={isVLAN}
     />
   );
 
-  const ActionMenu = flags.cmr ? LinodeActionMenu_CMR : LinodeActionMenu;
-
   return (
-    <React.Fragment>
-      {loading && (
-        <LinodeRowLoading
-          linodeStatus={status}
-          linodeId={id}
-          linodeRecentEvent={recentEvent}
-        >
-          {headCell}
-        </LinodeRowLoading>
-      )}
-      <TableRow
-        key={id}
-        className={classes.bodyRow}
-        data-qa-loading
-        data-qa-linode={label}
-        rowLink={`/linodes/${id}`}
-        ariaLabel={label}
+    <TableRow
+      key={id}
+      className={classes.bodyRow}
+      data-qa-loading
+      data-qa-linode={label}
+      rowLink={`/linodes/${id}`}
+      ariaLabel={label}
+    >
+      {headCell}
+      <TableCell
+        className={classNames({
+          [classes.statusCell]: true,
+          [classes.statusCellMaintenance]: maintenanceStartTime,
+          [classes.vlan_Status]: isVLAN
+        })}
+        data-qa-status
       >
-        {!loading && headCell}
-        <TableCell
-          parentColumn="Status"
-          className={classNames({
-            [classes.statusCell]: true,
-            [classes.statusCellMaintenance]: maintenanceStartTime
-          })}
-          data-qa-status
-        >
-          {!maintenanceStartTime ? (
-            loading ? (
-              'Busy'
-            ) : (
-              capitalize(displayStatus)
-            )
+        {!maintenanceStartTime ? (
+          loading ? (
+            <>
+              <StatusIcon status={iconStatus} />
+              <button
+                className={classes.statusLink}
+                onClick={() => openNotificationDrawer()}
+              >
+                <ProgressDisplay
+                  className={classes.progressDisplay}
+                  progress={getProgressOrDefault(recentEvent)}
+                  text={transitionText(status, id, recentEvent)}
+                />
+              </button>
+            </>
           ) : (
             <>
-              <div>
-                <div>
-                  <strong>Maintenance Scheduled</strong>
-                </div>
-                <div>
-                  {linodeMaintenanceWindowString(dateTime[0], dateTime[1])}
-                </div>
-              </div>
-              <HelpIcon
-                text={<MaintenanceText />}
-                className={classes.statusHelpIcon}
-                tooltipPosition="top"
-                interactive
-              />
+              <StatusIcon status={iconStatus} />
+              {displayStatus.includes('_')
+                ? capitalizeAllWords(displayStatus.replace('_', ' '))
+                : capitalize(displayStatus)}
             </>
-          )}
-        </TableCell>
-        <TableCell
-          parentColumn="IP Address"
-          className={classes.ipCell}
-          data-qa-ips
-        >
-          <div className={classes.ipCellWrapper}>
-            <IPAddress ips={ipv4} copyRight showCopyOnHover />
-          </div>
-        </TableCell>
-        <TableCell
-          parentColumn="Region"
-          className={classes.regionCell}
-          data-qa-region
-        >
-          <RegionIndicator region={region} />
-        </TableCell>
-        <LinodeRowBackupCell
-          linodeId={id}
-          backupsEnabled={backups.enabled || false}
-          mostRecentBackup={mostRecentBackup || ''}
-        />
-        <TableCell className={classes.actionCell} data-qa-notifications>
-          <div className={classes.actionInner}>
-            <RenderFlag
-              mutationAvailable={mutationAvailable}
-              linodeNotifications={linodeNotifications}
-              classes={classes}
+          )
+        ) : (
+          <div className={classes.maintenanceOuter}>
+            <strong>Maintenance Scheduled</strong>
+            <HelpIcon
+              text={<MaintenanceText />}
+              className={classes.statusHelpIcon}
+              tooltipPosition="top"
+              interactive
             />
-            <ActionMenu
+          </div>
+        )}
+      </TableCell>
+      {props.isVLAN ? (
+        <TableCell className={classes.ipCell} data-qa-ips>
+          <div className={classes.ipCellWrapper}>{vlanIP}</div>
+        </TableCell>
+      ) : null}
+      {props.isVLAN ? null : (
+        <>
+          <Hidden xsDown>
+            <TableCell className={classes.ipCell} data-qa-ips>
+              <div className={classes.ipCellWrapper}>
+                <IPAddress ips={ipv4} copyRight />
+              </div>
+            </TableCell>
+            <Hidden smDown>
+              <TableCell className={classes.regionCell} data-qa-region>
+                <RegionIndicator region={region} />
+              </TableCell>
+            </Hidden>
+          </Hidden>
+          <Hidden mdDown>
+            <LinodeRowBackupCell
               linodeId={id}
-              linodeLabel={label}
-              linodeRegion={region}
-              linodeType={type}
-              linodeStatus={status}
-              linodeBackups={backups}
-              openDeleteDialog={openDeleteDialog}
-              openPowerActionDialog={openPowerActionDialog}
-              openDialog={openDialog}
-              noImage={!image}
+              backupsEnabled={backups.enabled || false}
+              mostRecentBackup={mostRecentBackup || ''}
             />
-          </div>
-        </TableCell>
-      </TableRow>
-    </React.Fragment>
+          </Hidden>
+        </>
+      )}
+
+      <TableCell className={classes.actionCell} data-qa-notifications>
+        <div className={classes.actionInner}>
+          <RenderFlag
+            mutationAvailable={mutationAvailable}
+            linodeNotifications={linodeNotifications}
+            classes={classes}
+          />
+          <LinodeActionMenu
+            linodeId={id}
+            linodeLabel={label}
+            linodeRegion={region}
+            linodeType={type}
+            linodeStatus={status}
+            linodeBackups={backups}
+            openDialog={openDialog}
+            openPowerActionDialog={openPowerActionDialog}
+            noImage={!image}
+            inTableContext
+            inVLANContext={isVLAN}
+          />
+        </div>
+      </TableCell>
+    </TableRow>
   );
 };
 
@@ -247,7 +273,8 @@ const enhanced = compose<CombinedProps, Props>(
   withRecentEvent,
   hasMutationAvailable,
   withNotifications,
-  styled
+  styled,
+  React.memo
 );
 
 export default enhanced(LinodeRow);
@@ -273,6 +300,7 @@ export const RenderFlag: React.FC<{
   }
   if (linodeNotifications.length > 0) {
     return (
+      // eslint-disable-next-line
       <>
         {linodeNotifications.map((notification, idx) => (
           <Tooltip key={idx} title={notification.message}>
@@ -286,3 +314,18 @@ export const RenderFlag: React.FC<{
 };
 
 RenderFlag.displayName = `RenderFlag`;
+
+export const ProgressDisplay: React.FC<{
+  className?: string;
+  progress: null | number;
+  text: string;
+}> = props => {
+  const { progress, text, className } = props;
+  const displayProgress = progress ? `${progress}%` : `scheduled`;
+
+  return (
+    <Typography variant="body2" className={className}>
+      {text} {displayProgress === 'scheduled' ? '(0%)' : `(${displayProgress})`}
+    </Typography>
+  );
+};
