+++ packages/manager/src/features/linodes/LinodesLanding/LinodeRow/index.ts
@@ -1,2 +1,1 @@
-import LinodeRow from './LinodeRow';
-export default LinodeRow;
+export { default, ProgressDisplay, RenderFlag } from './LinodeRow';
